+++ webpack.prod.js
@@ -5,14 +5,14 @@ module.exports = {
 	devtool: 'source-map',	
 	mode: 'production',	
 	entry: {
-		'main.js' : "./src/main.js",
-		'config.js': "./src/config.js"
+		'main' : "./src/main.js",
+		'config': "./src/config.js"
 	},
 
 	output: {
-		path: path.resolve(__dirname, "./js"),
-    		filename: "[name]"
-  	},
+		path: path.resolve(__dirname, "js"),
+    		 filename: "[name].js"
+	  },
 
 	node: {
 		fs: 'empty'